Added Product Delete Role Security

+++ clear_sales.php
@@ -1,5 +1,18 @@
 <?php
 
+session_start();
+
+if ($_SERVER["REQUEST_METHOD"] !== "POST") {
+    die("Invalid Request");
+}
+
+if (
+    !isset($_SESSION["role"]) ||
+    $_SESSION["role"] !== "admin"
+) {
+    die("Access Denied");
+}
+
 include "db.php";
 
 $sql = "TRUNCATE TABLE sales";
